refactor: update structure for auth
