+++ client/app/(account)/(auth)/sign-in/page.tsx
@@ -1,0 +1,34 @@
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from '@/components/ui/card'
+import Link from 'next/link'
+import { LoginForm } from './login-form'
+
+export default function SignInPage() {
+  return (
+    <Card className="w-full max-w-md">
+      <CardHeader>
+        <CardTitle className="text-2xl">Sign in</CardTitle>
+        <CardDescription>
+          Enter your username below to login to your account
+        </CardDescription>
+      </CardHeader>
+      <CardContent>
+        <LoginForm />
+        <div className="mt-4 text-center text-sm">
+          Don&apos;t have an account?{' '}
+          <Link
+            href="/sign-up"
+            className="ml-1 text-base text-primary underline underline-offset-4"
+          >
+            Sign up
+          </Link>
+        </div>
+      </CardContent>
+    </Card>
+  )
+}
