+++ client/app/(account)/(auth)/sign-up/page.tsx
@@ -1,0 +1,34 @@
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from '@/components/ui/card'
+import Link from 'next/link'
+import { RegisterForm } from './register-form'
+
+export default function SignUpPage() {
+  return (
+    <Card className="w-full max-w-md">
+      <CardHeader>
+        <CardTitle className="text-xl">Sign Up</CardTitle>
+        <CardDescription>
+          Enter your information to create an account
+        </CardDescription>
+      </CardHeader>
+      <CardContent>
+        <RegisterForm />
+        <div className="mt-4 text-center text-sm">
+          Already have an account?{' '}
+          <Link
+            href="/sign-in"
+            className="ml-1 text-base text-primary underline underline-offset-4"
+          >
+            Sign in
+          </Link>
+        </div>
+      </CardContent>
+    </Card>
+  )
+}
