+++ client/package.json
@@ -9,6 +9,8 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@hookform/resolvers": "^3.3.4",
+    "@next/env": "^14.2.3",
     "@radix-ui/react-aspect-ratio": "^1.0.3",
     "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-dropdown-menu": "^2.0.6",
@@ -16,8 +18,7 @@
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-popover": "^1.0.7",
     "@radix-ui/react-scroll-area": "^1.0.5",
-    "@hookform/resolvers": "^3.3.4",
-    "@next/env": "^14.2.3",
+    "@radix-ui/react-separator": "^1.0.3",
     "@radix-ui/react-slot": "^1.0.2",
     "axios": "^1.6.8",
     "class-variance-authority": "^0.7.0",
@@ -28,12 +29,12 @@
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.51.4",
-    "sonner": "^1.4.41",
-    "tailwind-merge": "^2.2.2",
+    "react-infinite-scroll-component": "^6.1.0",
     "react-markdown": "^9.0.1",
     "remark-gfm": "^4.0.0",
+    "sonner": "^1.4.41",
     "swr": "^2.2.5",
-    "react-infinite-scroll-component": "^6.1.0",
+    "tailwind-merge": "^2.2.2",
     "zod": "^3.23.8"
   },
   "devDependencies": {
